requirements.txt
apache-airflow>=2.7.0
numpy==1.23.5
pandas==1.5.3
sqlalchemy==1.4.46
psycopg2-binary>=2.9.0
openpyxl>=3.1.0
